requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
django-filter>=23.0
python-dateutil>=2.8.0
